refactor: move limit orders widget actions to hook
