+++ src/cow-react/modules/limitOrders/containers/LimitOrdersWidget/index.tsx
@@ -1,22 +1,20 @@
 import * as styledEl from './styled'
 import { Field } from 'state/swap/actions'
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import React, { useMemo, useState } from 'react'
 import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/types'
 import { useFillLimitOrdersFullState, useLimitOrdersFullState } from '../../hooks/useLimitOrdersFullState'
-import { limitOrdersAtom, updateLimitOrdersAtom } from '../../state/limitOrdersAtom'
+import { updateLimitOrdersAtom } from '../../state/limitOrdersAtom'
 import { useAtomValue, useUpdateAtom } from 'jotai/utils'
 import { SettingsWidget } from '../SettingsWidget'
 import { limitOrdersSettingsAtom } from '../../state/limitOrdersSettingsAtom'
 import { RateInput } from '../RateInput'
 import { DeadlineInput } from '../DeadlineInput'
-import { useUpdateCurrencyAmount } from '../../hooks/useUpdateCurrencyAmount'
 import { LimitOrdersConfirmModal } from '../LimitOrdersConfirmModal'
 import { useTradeFlowContext } from '../../hooks/useTradeFlowContext'
 import { useIsSellOrder } from '../../hooks/useIsSellOrder'
 import { TradeButtons } from '@cow/modules/limitOrders/containers/TradeButtons'
 import { TradeApproveWidget } from '@cow/common/containers/TradeApprove/TradeApproveWidget'
 import { useSetupTradeState } from '@cow/modules/trade'
-import { useTradeNavigate } from '@cow/modules/trade/hooks/useTradeNavigate'
 import { ImportTokenModal } from '@cow/modules/trade/containers/ImportTokenModal'
 import { useOnImportDismiss } from '@cow/modules/trade/hooks/useOnImportDismiss'
 import { limitRateAtom } from '../../state/limitRateAtom'
@@ -28,9 +26,6 @@ import { OrderKind } from '@cowprotocol/cow-sdk'
 import { useWalletInfo } from '@cow/modules/wallet'
 import { useDetectNativeToken } from '@cow/modules/swap/hooks/useDetectNativeToken'
 import { LimitOrdersProps, limitOrdersPropsChecker } from './limitOrdersPropsChecker'
-import tryParseCurrencyAmount from 'lib/utils/tryParseCurrencyAmount'
-import { useOnCurrencySelection } from '@cow/modules/trade/hooks/useOnCurrencySelection'
-import { FractionUtils } from '@cow/utils/fractionUtils'
 import { useSetupLimitOrderAmountsFromUrl } from '@cow/modules/limitOrders/hooks/useSetupLimitOrderAmountsFromUrl'
 import { InfoBanner } from '@cow/modules/limitOrders/pure/InfoBanner'
 import { partiallyFillableOverrideAtom } from '@cow/modules/limitOrders/state/partiallyFillableOverride'
@@ -39,6 +34,7 @@ import { useFeatureFlags } from '@cow/common/hooks/useFeatureFlags'
 import { TradeWidget } from '@cow/modules/trade/containers/TradeWidget'
 import usePriceImpact from '@src/hooks/usePriceImpact'
 import { useRateInfoParams } from '@cow/common/hooks/useRateInfoParams'
+import { useLimitOrdersWidgetActions } from '@cow/modules/limitOrders/containers/LimitOrdersWidget/hooks/useLimitOrdersWidgetActions'
 
 export function LimitOrdersWidget() {
   useSetupTradeState()
@@ -60,20 +56,16 @@ export function LimitOrdersWidget() {
     isUnlocked,
     orderKind,
   } = useLimitOrdersFullState()
-  const onCurrencySelection = useOnCurrencySelection()
   const onImportDismiss = useOnImportDismiss()
-  const limitOrdersNavigate = useTradeNavigate()
   const settingsState = useAtomValue(limitOrdersSettingsAtom)
-  const updateCurrencyAmount = useUpdateCurrencyAmount()
   const isSellOrder = useIsSellOrder()
   const tradeContext = useTradeFlowContext()
-  const state = useAtomValue(limitOrdersAtom)
-  const updateLimitOrdersState = useUpdateAtom(updateLimitOrdersAtom)
-  const { isLoading: isRateLoading, activeRate, feeAmount } = useAtomValue(limitRateAtom)
+  const { isLoading: isRateLoading, feeAmount } = useAtomValue(limitRateAtom)
   const rateInfoParams = useRateInfoParams(inputCurrencyAmount, outputCurrencyAmount)
   const { isWrapOrUnwrap } = useDetectNativeToken()
   const partiallyFillableOverride = useAtom(partiallyFillableOverrideAtom)
   const { partialFillsEnabled } = useFeatureFlags()
+  const widgetActions = useLimitOrdersWidgetActions()
 
   const showRecipient = useMemo(
     () => !isWrapOrUnwrap && settingsState.showRecipient,
@@ -107,53 +99,6 @@ export function LimitOrdersWidget() {
     fiatAmount: outputCurrencyFiatAmount,
     receiveAmountInfo: null,
   }
-  const onUserInput = useCallback(
-    (field: Field, typedValue: string) => {
-      const currency = field === Field.INPUT ? inputCurrency : outputCurrency
-
-      if (!currency) return
-
-      const value = tryParseCurrencyAmount(typedValue, currency) || null
-
-      updateCurrencyAmount({
-        activeRate,
-        amount: value,
-        orderKind: isWrapOrUnwrap || field === Field.INPUT ? OrderKind.SELL : OrderKind.BUY,
-      })
-    },
-    [updateCurrencyAmount, isWrapOrUnwrap, inputCurrency, outputCurrency, activeRate]
-  )
-
-  const onSwitchTokens = useCallback(() => {
-    const { inputCurrencyId, outputCurrencyId } = state
-
-    if (!isWrapOrUnwrap) {
-      updateLimitOrdersState({
-        inputCurrencyId: outputCurrencyId,
-        outputCurrencyId: inputCurrencyId,
-        inputCurrencyAmount: FractionUtils.serializeFractionToJSON(outputCurrencyAmount),
-        outputCurrencyAmount: FractionUtils.serializeFractionToJSON(inputCurrencyAmount),
-        orderKind: orderKind === OrderKind.SELL ? OrderKind.BUY : OrderKind.SELL,
-      })
-    }
-    limitOrdersNavigate(chainId, { inputCurrencyId: outputCurrencyId, outputCurrencyId: inputCurrencyId })
-  }, [
-    state,
-    isWrapOrUnwrap,
-    limitOrdersNavigate,
-    updateLimitOrdersState,
-    chainId,
-    inputCurrencyAmount,
-    outputCurrencyAmount,
-    orderKind,
-  ])
-
-  const onChangeRecipient = useCallback(
-    (recipient: string | null) => {
-      updateLimitOrdersState({ recipient })
-    },
-    [updateLimitOrdersState]
-  )
 
   const props: LimitOrdersProps = {
     inputCurrencyInfo,
@@ -165,10 +110,6 @@ export function LimitOrdersWidget() {
     isExpertMode,
     recipient,
     chainId,
-    onChangeRecipient,
-    onUserInput,
-    onSwitchTokens,
-    onCurrencySelection,
     onImportDismiss,
     partiallyFillableOverride,
     featurePartialFillsEnabled: partialFillsEnabled,
@@ -177,15 +118,8 @@ export function LimitOrdersWidget() {
     tradeContext,
     settingsState,
     feeAmount,
+    widgetActions,
   }
-
-  /**
-   * Reset recipient value only once at App start
-   */
-  useEffect(() => {
-    onChangeRecipient(null)
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
 
   return <LimitOrders {...props} />
 }
@@ -197,9 +131,7 @@ const LimitOrders = React.memo((props: LimitOrdersProps) => {
     isUnlocked,
     chainId,
     isRateLoading,
-    onUserInput,
-    onSwitchTokens,
-    onCurrencySelection,
+    widgetActions,
     onImportDismiss,
     partiallyFillableOverride,
     featurePartialFillsEnabled,
@@ -207,7 +139,6 @@ const LimitOrders = React.memo((props: LimitOrdersProps) => {
     showRecipient,
     isExpertMode,
     recipient,
-    onChangeRecipient,
     rateInfoParams,
     priceImpact,
     tradeContext,
@@ -274,13 +205,6 @@ const LimitOrders = React.memo((props: LimitOrdersProps) => {
     ),
   }
 
-  const actions = {
-    onCurrencySelection,
-    onUserInput,
-    onChangeRecipient,
-    onSwitchTokens,
-  }
-
   const params = {
     disableNonToken: false,
     compactView: false,
@@ -295,7 +219,7 @@ const LimitOrders = React.memo((props: LimitOrdersProps) => {
     <>
       <TradeWidget
         slots={slots}
-        actions={actions}
+        actions={widgetActions}
         params={params}
         inputCurrencyInfo={inputCurrencyInfo}
         outputCurrencyInfo={outputCurrencyInfo}
