+++ src/cow-react/modules/trade/containers/TradeWidget/index.tsx
@@ -2,7 +2,7 @@ import * as styledEl from './styled'
 import { TradeWidgetLinks } from '@cow/modules/application/containers/TradeWidgetLinks'
 import { CurrencyInputPanel, CurrencyInputPanelProps } from '@cow/common/pure/CurrencyInputPanel'
 import { CurrencyArrowSeparator } from '@cow/common/pure/CurrencyArrowSeparator'
-import React from 'react'
+import React, { useEffect } from 'react'
 import { useDetectNativeToken } from '@cow/modules/swap/hooks/useDetectNativeToken'
 import { useWalletDetails, useWalletInfo } from '@cow/modules/wallet'
 import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/types'
@@ -12,7 +12,7 @@ import { PriceImpact } from 'hooks/usePriceImpact'
 import { SetRecipientProps } from '@cow/modules/swap/containers/SetRecipient'
 import { t } from '@lingui/macro'
 
-interface TradeWidgetActions {
+export interface TradeWidgetActions {
   onCurrencySelection: CurrencyInputPanelProps['onCurrencySelection']
   onUserInput: CurrencyInputPanelProps['onUserInput']
   onChangeRecipient: SetRecipientProps['onChangeRecipient']
@@ -76,6 +76,14 @@ export function TradeWidget(props: TradeWidgetProps) {
 
   // Disable too frequent tokens switching
   const throttledOnSwitchTokens = useThrottleFn(onSwitchTokens, 500)
+
+  /**
+   * Reset recipient value only once at App start
+   */
+  useEffect(() => {
+    onChangeRecipient(null)
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
 
   return (
     <styledEl.Container id={id}>
